+++ db/migrate/20000101000000_enable_agilibox_extensions.rb
@@ -1,0 +1,7 @@
+class EnableAgiliboxExtensions < ActiveRecord::Migration[5.1]
+  def change
+    enable_extension "pgcrypto"
+    enable_extension "unaccent"
+    enable_extension "uuid-ossp"
+  end
+end