+++ db/migrate/20170502143330_enable_unaccent.rb
@@ -1,5 +1,0 @@
-class EnableUnaccent < ActiveRecord::Migration[5.0]
-  def change
-    enable_extension "unaccent"
-  end
-end