+++ spec/dummy/db/schema.rb
@@ -10,24 +10,26 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20170502143330) do
+ActiveRecord::Schema.define(version: 20170223093800) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
+  enable_extension "pgcrypto"
   enable_extension "unaccent"
+  enable_extension "uuid-ossp"
 
-  create_table "dummy_models", force: :cascade do |t|
-    t.string   "string_field"
-    t.text     "text_field"
-    t.integer  "integer_field"
-    t.decimal  "decimal_field"
-    t.boolean  "boolean_field"
-    t.date     "date_field"
+  create_table "dummy_models", id: :uuid, default: -> { "uuid_generate_v4()" }, force: :cascade do |t|
+    t.string "string_field"
+    t.text "text_field"
+    t.integer "integer_field"
+    t.decimal "decimal_field"
+    t.boolean "boolean_field"
+    t.date "date_field"
     t.datetime "datetime_field"
-    t.integer  "asso_id"
-    t.string   "asso_type"
-    t.datetime "created_at",     null: false
-    t.datetime "updated_at",     null: false
+    t.uuid "asso_id"
+    t.string "asso_type"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
   end
 
 end