+++ spec/libs/agilibox/collection_update_spec.rb
@@ -11,8 +11,11 @@ describe Agilibox::CollectionUpdate do
     end
   }
 
-  let!(:instance1) { model.create!(id: 1, string_field: "A") }
-  let!(:instance2) { model.create!(id: 2, string_field: "B") }
+  let(:uuid1) { Agilibox::SortableUUIDGenerator.call }
+  let(:uuid2) { Agilibox::SortableUUIDGenerator.call }
+
+  let!(:instance1) { model.create!(id: uuid1, string_field: "A") }
+  let!(:instance2) { model.create!(id: uuid2, string_field: "B") }
 
   def updater_for(arr)
     described_class.new(model.all, arr)
@@ -20,8 +23,8 @@ describe Agilibox::CollectionUpdate do
 
   it "#update should update objects and return true" do
     service = updater_for [
-      {id: 1, string_field: "X"},
-      {id: 2, string_field: "Y"},
+      {id: uuid1, string_field: "X"},
+      {id: uuid2, string_field: "Y"},
     ]
 
     expect(service.update).to eq true
@@ -31,8 +34,8 @@ describe Agilibox::CollectionUpdate do
 
   it "#update should NOT update any objects and return false in any object is invalid" do
     service = updater_for [
-      {id: 1, string_field: ""},
-      {id: 2, string_field: "Y"},
+      {id: uuid1, string_field: ""},
+      {id: uuid2, string_field: "Y"},
     ]
 
     expect(service.update).to eq false
@@ -42,8 +45,8 @@ describe Agilibox::CollectionUpdate do
 
   it "#update should destroy objects width _destroy=1" do
     service = updater_for [
-      {id: 1, _destroy: 0},
-      {id: 2, _destroy: 1},
+      {id: uuid1, _destroy: 0},
+      {id: uuid2, _destroy: 1},
     ]
 
     expect(service.update).to eq true
@@ -53,8 +56,8 @@ describe Agilibox::CollectionUpdate do
 
   it "#update! should raise exception on invalid record" do
     service = updater_for [
-      {id: 1, string_field: ""},
-      {id: 2, string_field: "Y"},
+      {id: uuid1, string_field: ""},
+      {id: uuid2, string_field: "Y"},
     ]
 
     expect { service.update! }.to raise_error(ActiveRecord::RecordInvalid)